archival: Fix high reupload rate
The backoff algorithm is triggered if the number of uploaded segments
is zero. This enables backoff in case there is nothing to upload or if
there are upload errors.

However, if segment uploads are discarded due to consistency violations,
the successful upload counter is not updated. If all uploads are
discarded, the backoff algorithm is not triggered, and the upload loop
behaves as if a segment was uploaded successfully. As a result, in the
presence of persistent consistency violations, ntp_archiver causes high
I/O utilization by repeatedly retrying the same uploads.

This commit fixes the issue by adjusting the successful upload counter
after performing metadata consistency checks.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>
(cherry picked from commit d7a5c0d618a1317b55e4a681189d8e93f0f4fd6f)