dt: add missing schema registry flags to acl list
The command used to be only used for Kafka, for
docker we don't need this as rpk will infer the
registry listener. But for CDT we need to
explicitly add the flags as we can't rely on
localhost.

(cherry picked from commit 66c4a9c3d2cc9652d39f9a1f318c5a77d0c71426)