cluster_link/task: correctly change task interval while task running
`task::runner::set_task_interval` rearmed the timer even if a task is
running. Both timer callback function and initial `start()` call
unconditionally rearm the timer after the task is complete. It may lead
to unsupported timer operations, such as double rearm and unarmed timer
timeout lookup. The solution is to skip rearming in
`task::runner::set_task_interval` if the timer is armed already.

I did not add a vassert to check the timer is not armed in the other two
places we arm it because such an assert is already in place in Seastar.
It only works in debug mode though.