c/leaders_table: wire in controller abort source
Controller input is stopped in the early stage of Redpanda shutdown
process. This way all the request are failed early when Redpanda is
shutting down. Passing in a `controller` abort source to the
`partition_leaders_table` will allow us to interrupt all fibers waiting
for the leaders instead of having to wait for the timeout to expire.

Signed-off-by: Michał Maślanka <michal@redpanda.com>