rpk cluster config: install --all flag locally
It was previously installed at the root of the
'rpk cluster config' command, resulting in
appearing in all subcommands' help text even if
those didn't use it. This was confusing, so now we
just install it 'locally' in the commands that use
the flag.

(cherry picked from commit 7a28d335ef689d40419f483a48bdaec61867508b)