tests: added test validating force recovery from local state
Added a test validating if force recovery works correctly when local
state of the one of the replicas is more up to date than the cloud
state. The test achieves that by not waiting for the manifest uploads
and reducing the frequency of cloud uploads.

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit 8deff1e3d3bb0e1e6685e956ceb2b3993d460a22)