ct: Move level_zero_log_reader back to frontend
The level_zero_log_reader is a part of the frontend logically. It
bridges the gap betwen the cluster and ct/data_plane_api. It was
extracted out of the frontend at some point.

The stuff in ct/level_zero contains low level pipeline components like
batcher and throttler. The 'reader' module contains pipeline components
used in the fetch path. These components are local to the L0 subsystem.

The level_zero_log_reader_impl depends on the code in the
ct/level_zero/reader but it's not its part. Also, the code in the
ct/level_zero shouldn't depend on cluster (except maybe the stm).

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>