rpk debug bundle: avoid storing large files on memory
This is to reduce the memory usage of rpk when
running debug bundle. We were storing the files
into memory, and some of them grow as the cluster
grows (e.g data directory), causing pressure on the
memory available in the cluster.

We solve this by :

1. When reading files: streaming the file directly
   into the Zip writer.

2. When constructing large datasets (kafka metadata
   or data-dir): by using buffered writers to
   write to temporary files and then stream into
   the zip writer.

3. Using the 'stream' option of the rpadmin package
   for the Metrics endpoints, as their response
   grows with the partition count in the cluster.

(cherry picked from commit 1ec79472db71f323b5efb70c8cf57551b37d6e06)