kafka/protocol: non-zero default value for `PartitionMaxBytes`
Previously the `max.partition.fetch.bytes` field in fetch requests was
being ignored in the fetch path. Therefore using the default field value
of 0 still resulted in data being returned for the partition. However, the
field is now being correctly used in the fetch path and setting it to 0 will
result in no bytes being returned for that partition.

This broke the expectations of many tests that didn't set
`PartitionMaxBytes` and still expected data to be returned for that
partition. Hence this commit changes the default value to be non-zero.