admin: require controller leadership before license idempotence check
Ensure that PUT /v1/features/license runs on the controller leader
before performing the idempotence check. Without this, a follower that
is slightly behind could incorrectly return 200 OK if the license
history went A -> B -> A, since it would see the last request as
idempotent when it was not.

This surfaced as CI flakiness in a test exercising the A -> B -> A case.
In practice, customers are unlikely to encounter it because license
updates are infrequent, but it is still safer to perform the idempotence
check on the leader.

(cherry picked from commit 1144acd3a496f97d763f23c184de54ea8044b18f)