dt: check for license on all nodes
The test_license_upload_and_query test changes the license from v0 -> v1
-> v0. This has been flaky because of the idempotence check on stale
nodes (see the previous commit's message).

The previous commit ensures that the license update runs on the leader
in the "happy case" of when leadership is stable. However, in the face
of unstable leadership, this test could still flake because a stale node
could still think that it is the leader and fall into the same
idempotence scenario described earlier. To avoid flakiness in this case
as well, i.e. to make the test robust to leadership transfers, this
commit ensures that the license is checked on all nodes instead of any
random node after each update.

(cherry picked from commit de59b5202df8627d5c960ed4a25141d3f98b72d5)