tests/data_migrations_api_test: deflake verification
KgoProducer cannot stop with waiting for acks.
When it is blocked by migrations let it get stuck for a while to gather
acks. When it is not limit the number of messages produced so the number
of acked messages can be determined.

(cherry picked from commit 3ee71fe84f19b69930efaaa06c1ec370058f97a7)