archival: adjacent_segment_merger should not get stuck on a bad run
Adjacent segment merging only runs when compaction is disabled, but if
compaction was previous enabled for some topic then there may be compacted
segments on disk for a while (or indefinitely depending on the retention
policy).

If these segments:
- Were uploaded to cloud storage before compaction
- Were NOT reuploaded after compaction
- Were originally small enough to be mergeable relative to the configured
  remote segment target size.

Then the adjacent segment merger can get in a nasty situation where it scans
the partition manifest for a run of small segments and proposes that run for
merging but the sizes recorded in the partition manifest don't match the
actual (compacted) size of the segments on disk. Segment collection can
succeed but the content size of the resulting upload candidate doesn't match
the run and so is rejected by the housekeeping job.

Since no upload took place, the '_last' offset pointer in the housekeeping job
is not updated, so the merger attempts to reupload the same, incorrectly sized
run effectively forever (or until the offending compacted segments are removed
from the local log).

This commit introduces 'find_reupload_candidate_result::skip_to', an optional
rp offset that indicates to the housekeeping job that, although no upload
took place, the '_last' pointer should still be updated so that the merger
can make progress.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>
(cherry picked from commit c994a11a360016346e8c7869b6ceac066228aa26)