bazel/packaging: allow specifying install path as a prefix
When building Redpanda and direct consumer verifier packages the only
component of interpreter path that changes is the package directory
itself. Previously it was required to provide the whole install path
when building packages. This lead to bazel configuration change and
unnecessary rebuild.

This PR changes the semantics of `install_path` so that it only
specifies the install path up to the package directory.

The directory itself is configured separately. This way two packages
sharing the same install path root may be build using a single bazel
command.

Signed-off-by: Michał Maślanka <michal@redpanda.com>