kafka fetch: make planning hierarchical
Current fetch planning iterates over all partitions in the fetch
request (or fetch session, if sessions are used) in a "flat" manner,
i.e., {"topic1", 0}, {"topic1", 1}, {"topic2", 0} for 3 partitions
spread across two topics. Many of the checks in the planning loop are
topic-specific, i.e,. they depend on only the topic, not the partition.

In the current flat approach, these checks are repeatedly executed for
the same topic in the case of wide reads with multiple partitions in
the same topic (the most common case for wide reads).

In this series we change the processing to be "hierarchical" instead,
i.e., nested loops with topic on the outside, partition in the inside.
When processing a request directly (i.e., not fetch session) the input
is already hierarchical in this way. When iterating over a fetch
session, it is not, so we re-group the flat list by topic at iteration
time. In the session case, a topic could still appear more than once
since we do an "in order" grouping and there is no guarantee that
partitions for a given topic are all consecutive in session partition
order.