config: add sasl_mechanisms_overrides config
This new config is a way to control the cluster's sasls_mechanisms
on a per-listener basis. For a given kafka listener, if there is
an entry in the sasl_mechanisms_overrides config, the value from
sasl_mechanisms will be ignored and the override will be used in its
place.

For any functionality that needs to know if a sasl mechanisms is active,
like for the enterprise features check, it is considered active if it
is used in any listener override.

(cherry picked from commit 5a790842b022a642d762213fa1e9a04b6094f798)