c/a/tests: Add segment_collector_rpbench
Fairly basic benchmark comparing the two extant methods of producing an
upload candidate. Since the async_data_uploader-based version is neither in
use nor optimal, it's not a very interesting benchmark as written. The idea
here is to introduce a baseline as we swap out the guts of the implementation
to remove duplicated work.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>