ct: add support for prefix_truncate
While largely lifted from replicated_partition, prefix truncation will
work similar to the way it does today for tiered storage, the difference
in applying will be truncation, which in cloud topics will be done by
the reconciler. The reconciler will have access to the log eviction STM
so it can apply the changes there. Since there are in flight changes
to the reconciler, these changes are only for L0.