iceberg/rest_client: log more about request on error
CONFLICT:
- minor conflict in rest_client/BUILD

When there is an error returned from the REST catalog, we have very
little context about what we sent. This commit addresses this by
logging the logical contents of the request as JSON.

I considered logging at a lower level, and logging the entire HTTP
request payload (e.g. in perform_request() after signing and such), but
opted to place the at the callsite because the context we have there is
already const and available to be logged conditionally (vs in
perform_request(), the request is moved, making conditional logging
trickier).

(cherry picked from commit e315237b7434131ceb1ad3bb34995e11769c3592)