c/producer_seq: accept batches with last seq < than first seq
Batch identity with `fist_seq` > `last_seq` is a valid batch if an
overflow in sequence number occurred. In this case a first sequence will
be a large number (close to `int32_t` max value) while the last sequence
will be wrapped around and difference will be added to `0`.

This is a valid situation and producer should still be able to continue.
Returning an `invalid_request` error lead to a situation in which
producer can not continue and partition leader is stepping down as an
error is interpreted as an replication error by the `rm_stm`

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit 624ae4a8e1931dd3d8963725a5e5b08c6a5f0e3b)