ts: Improve flush operation
The flush could potentially take longer if there is a concurrent scrub
or adjacent segment merging run or housekeeping is going to run.

This commit improves flush latency a bit by disabling scrubbing when the
flush is invoked.

If the archiver uploads segments normally the flush will have to wait
until ongoing uploads complete and then the housekeeping may also run.
Only after that the flush could be completed by uploading the rest of
the partition. This commit disables this housekeeping run so the flush
could be completed earlier (maybe_complete_flush is invoked without
invoking housekeeping).

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>