kafka/server: change bytes_left calculation in fill_fetch_responses
Changes the `bytes_left` calculation when determining if there is
enough space for a given partition response. Previously
`octx.bytes_left` was used. However, this value includes the size of
the last read result written to the response for the partition. Hence
that pervious result needs to be subtracted out first to get the
accurate number of bytes left.