c/producer_state: explicitly handle negative batch sequence numbers
Batch sequence can never be negative when producer is idempotent. It is
required to start from '0' and it has a special handling for the
overflow. Added explicit handling of negative batch sequence numbers as
the other parts of producer state logic depends on the sequence being a
number greater than or equal to `0`

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit 2177b1f1079035efb7da2751e317443487e69b0c)