c/tx: fix begin of transactions without a fence batch
If a fence batch is truncated, we restart from the first available
transaction batch. This fix addresses a bug where that logic fails
when the producer has a previously committed or aborted transaction.

For each producer, we track either an ongoing transaction or the
last completed one. The logic did not account for completed
transactions.

When a finished transaction exists and a new transactional batch
arrives, we must start a new transaction—but this was not handled
correctly.

(cherry picked from commit 66f5f72e58627551e56b0bf863e5ddad292aacea)