ts: Add hydration failsafe
The remote segment has few hydration methods that download objects from
cloud storage and putting them to the cloud storage cache:
- do_hydrate_txrange
- do_hydrate_index
- do_hydarte_segment
- hydrate_chunk

The last method checks the cache before performing the download but the
first three doesn't. This commit adds similar check to first three
methods (check if the object is in the cache and only if it's not
download it and put to the cache).