`ssx`: support different `Rate` types in `rate_limited_function`
The inspiration here is to be able to support more complex types than
just a bare duration type within the `rate_limited_function`. For example,
for a duration type `T`, a motivating case would be a `binding<T>` to a
cluster property controlling an interval rate:

```
auto fn = ssx::rate_limited_function<
   void,
   seastar::lowres_clock,
   config::binding<std::chrono::milliseconds>>(
   [this] {
       foo();
   },
   config::shard_local_cfg().foo_interval.bind());
```

or an `optional`:

```
auto fn = ssx::rate_limited_function<
   void,
   seastar::lowres_clock,
   std::optional<std::chrono::milliseconds>>(
   [this] {
       foo();
   },
   config::shard_local_cfg().optional_foo_interval()),
```

or a `binding<std::optional<T>>`, or even something like a
`ss::noncopyable_function<T()`>, and so on and so forth.

The changes in this commit allow for these more complex types to be used
within the `rate_limited_function`.