kafka/server: change obligatory batch read behavior
This commit changes the obligatory batch reads to first do a non-waiting
get of memory units. Followed by a read of the partition up to the
number of acquired units in size. If that fails to read a batch then it
waits for enough units to read any batch from the partition.