rm_stm: do not reinsert producer to active tx producers list if present
Recently we relaxed an assertion in the producer state. The assertion
trigger when control batch responsible for ending transaction was lost.
This should be treated as an error but it is possible as a result of
force recovery operation or other situation in which we accept a data
loss.

This PR adds a check verifying if producer is already in the active
transactional producers list when beginning transaction.

Previously without the check assertion was triggered.

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit 9197536df66d885eda7a4c982cff7a0d51f5a51b)