ct: move shutdown to after partition shutdown
In the many partitions test with cloud topics, Redpanda would
occasionally hang when shutting down. This was because we would wait for
a Raft replication attempt to finish when shutting down the L1 leader
router, but it wouldn't complete with the RPC subsystem shutdown already
happening.

The solution is to shut down the partitions first, which in turn stops
further Raft replication. This is the suggested solution for other STMs,
as indicated by nearby comments in application.cc.