tools/dev_cluster: disable grafana/prometheus by default
99% of the time these aren't useful. I often re-create the dev
environment (rm -rf data) which leads to long startup times because of
grafana. I already run grafana instance on my local machine which is
bound to default port (default apt pkg config) so I have to do extra
customizations for simple dev cluster to work.

Flip the defaults so that they work for 99% of the cases. In the 1% of
the cases when monitoring is desired pass the flags.