partition_balancer: report quorum loss moves as...
immutable

nodes_decommissioning_test.py
::NodesDecommissioningTest
.test_decommissioning_node_rf_1_replica

would periodically fail on partitions not being reported as allocation
failures. This happened because there was a race. A partition would NOT
be reported as an allocation failure if there was a move in progress.

In this test, the node is stopped and then the decommed. As a result,
the broker could be picked up as unresponsive, which would init a move
before the decomission is made visible to the planner.

This commit changes pbp to report in-progress moves with quorum loss on
the original replica set as immutable.

(cherry picked from commit 1ac4c69f888db5698f323543aa506df21045318a)