shared_placement_scale_test: adjust threshold
Adjust the expected throughput threshold.

Empirically this test gets the expected throughput (100 MiB/s), but this
is not an exact measurement given sampling intervals, distributed
clients which start and stop and operated in a batched manner, so we
can't really just compare that the observed throughput is == or >= the
target throughput, since this will flake, rather we check that is above
the given fraction of the nominal/expected throughput.

The observed throughput in the cases I looked at was much closer, e.g.,
between 99 and 101% of the target, though I saw as low as 97% in one
test out of ~15.

After this change 39 of 40 runs passed and the other failure wasn't
throughput related.

(cherry picked from commit 8cdec917f2963d74c47784fbdebb47324a13ef1c)