rptest: Allow rpc error log line
The log line is:
pc_server.cc:263 - Service handler for method 3105104740 threw an
exception: std::runtime_error (ntp {kafka/topic/0}: log offset 0
is outside the translation range (starting at 5048))

This is expected and can happen when log truncation races with partition
movement. The operation is retried and the test passes.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>