rptest: Enable fast uploads in datalake test
The test expects uplads to happen in 10 minutes. The problem is that it
create different number of partitions in docker and on a dedicated
machine (100 vs 1000). With 1000 partitions the chance that one
partition will receive not enough data to roll a segment is relatively
high.

The fix is to force uploads and to increase the amount of data on a
dedicated node.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>
(cherry picked from commit e8e0a175424ab4df39d5894bf2fb921c632d071c)