gtest: give stacktraces on sigill
Use absl's signal handler in gtest in order to give proper backtraces.
We don't use crash_tracker because it's undesirable to bring in a bunch
of dependencies for such a low level thing. If someone wanted to extract
out just the backtrace stuff for tests from crash_tracker, I'd accept
it, but for now this will do.