kafka: return partition error metadata if no term
When the metadata cache is missing term information for a given partition (e.g.
when starting up) we return an invalid (-1) term with no error.

The Java client treats this metadata as unreliable and forgets about any cached
epochs, regardless of any errors[1].

Franz go skips processing the partition altogether if there is an error,
regardless of whether there is a missing term, opting to retry later. It
otherwise accepts the missing term (-1) as a valid term and treating
it as a term moving backwards[2].

This updates the metadata handler to return an error, which seems like it
shouldn't affect the Java client, but should help Franz go ignore the invalid
metadata.

Relevant Redpanda commits in reverse chronological order:
- https://github.com/redpanda-data/redpanda/commit/a441290876ff9bb01d6d58532f388b94f3719bfe
- https://github.com/redpanda-data/redpanda/commit/7a60b758ad9ff9f2a2d7ec343f2a404ffaf51e89
- https://github.com/redpanda-data/redpanda/commit/86d583b486d033377be4c9f2b331be25ff6d9632

1. https://github.com/apache/kafka/blob/9529003fffd93a9d7e3f6ff7ab081ed84942fd13/clients/src/main/java/org/apache/kafka/clients/Metadata.java#L596-L601
2. https://github.com/twmb/franz-go/blob/3affad808a82ebfe1555d01d27732431093ac8e1/pkg/kgo/metadata.go#L811-L857