admin: Store decommission_status partitions in a chunked array
The in-memory form of the decommission_status API response object stored
the partitions in a vector. When there are a lot of partitions being
decommissioned, this vector could grow large enough to trigger an
oversized allocation warning when it resizes.

The fix is to use the chunked_array type instead.

(cherry picked from commit 3ceb74d2254d57e67319bf3b3f415dfacf905329)