schema_registry: merge swagger json
Store the schema registry swagger definition as a single json file.

Use a bazel helper to split the file into the components that the
seastar tooling requires to generate the C++ code.

This is required to be be able to auto-generated a python client for the
API, and it also makes changing/reading the swag nicer.