`storage`: reduce calls to feature table and cluster config... ...for checking enablement of tx batch removal.
It seems like there was previously the potential for a race between
enablement of this setting and compaction on the granularity of
a single segment (which would result in undefined behavior if we were
able to e.g. query `false` when attempting to remove a fence batch
but then querying `true` when attempting to remove a control batch),
it wasn't outright obvious from a glance at the code.

Rework this code so that calls to `is_tx_batch_compaction_enabled()` are
minimized in the storage layer- at most once per instanation of
`copy_data_segment_reducer` and `should_keep()` lambdas to be certain that
the above situation cannot happen.

The other locations which call to `is_tx_batch_compaction_enabled()` are
when considering if a `segment` needs a rewrite in sliding window compaction
(if this races with the cluster config being toggled, nothing bad could happen)
as well as calls to `maybe_rebuild_compaction_index()` (once again, no potential
pitfalls if this were to race).