test: cleanup redpanda service on preallocated nodes
preallocated nodes may be used by any service (e.g. redpanda broker, OMB
worker, etc..). However, note below in free_preallocated_nodes that
clean-up specific to a redpanda broker is performed
(RedpandaService::sockets_clear), which waits for Redpanda broker ports
to be cleaned up. We've seen that fail on a pre-allocated node which was
using a non-Redpanda broker service, suggesting that a previous user of
the node was a Redpanda broker and it failed to clean up, and that most
likely that broker was still running and listening on the port. So
before we use the preallocated node clean off the node w.r.t. the
Redpanda broker service even if it is not the service which will be this
node.

An example here is where we time out waiting for 33145 to be cleared,
which is redpanda internal rpc port. but node .71 is a OMB worker, so
most likely an old redpanda broker is still running.

[DEBUG - 2025-12-03 16:25:35,763 - redpanda - sockets_clear - lineno:3651]: node=ip-clipped-71 Active Internet connections (servers and established)
[DEBUG - 2025-12-03 16:25:35,763 - redpanda - sockets_clear - lineno:3651]: node=ip-clipped-71 Proto Recv-Q Send-Q Local Address           Foreign Address         State
[DEBUG - 2025-12-03 16:25:35,763 - redpanda - sockets_clear - lineno:3651]: node=ip-clipped-71 tcp        0      0 0.0.0.0:40487           0.0.0.0:*               LISTEN
[DEBUG - 2025-12-03 16:25:35,763 - redpanda - sockets_clear - lineno:3651]: node=ip-clipped-71 tcp        0      0 127.0.0.53:53           0.0.0.0:*               LISTEN
[DEBUG - 2025-12-03 16:25:35,763 - redpanda - sockets_clear - lineno:3651]: node=ip-clipped-71 tcp        0      0 0.0.0.0:9999            0.0.0.0:*               LISTEN
[DEBUG - 2025-12-03 16:25:35,763 - redpanda - sockets_clear - lineno:3651]: node=ip-clipped-71 tcp        0      0 0.0.0.0:9091            0.0.0.0:*               LISTEN
[DEBUG - 2025-12-03 16:25:35,763 - redpanda - sockets_clear - lineno:3651]: node=ip-clipped-71 tcp        0      0 0.0.0.0:9090            0.0.0.0:*               LISTEN
[DEBUG - 2025-12-03 16:25:35,763 - redpanda - sockets_clear - lineno:3651]: node=ip-clipped-71 tcp        0      0 0.0.0.0:33145           0.0.0.0:*               LISTEN
[INFO  - 2025-12-03 16:25:35,763 - redpanda - sockets_clear - lineno:3661]: Port collision on node ip-clipped-71: 0.0.0.0:33145->0.0.0.0:* LISTEN

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>