cluster/cl: Batch failover link commands
There is an issue when attempting a failover of the entire link when
there are thousands of shadow topics to fail over.  This is because we
submit a single update status command per shadow topic.  This takes some
time and additionally the way that shadow link metadata is copied can
also cause reactor stalls.  This results in the UX receiving a command
timeout.

Here, we will 'batch' the update command by using an empty topic name to
indicate a batch update.  THis is a special case _only_ for fail over.

Signed-off-by: Michael Boquard <michael@redpanda.com>