ct/l1/lsm: add state_reader
Adds an abstraction to get L1 metastore state from a database snapshot.
Long term, this abstraction will be used to serve the metastore
interface, as well as to help validate invariants when performing
metastore state updates.

It currently doesn't include all of the read paths required for serving
the metastore, but has enough to be used in an upcoming change to
implement invariant checks for add_objects and replace_objects.